fix relayer-register

+++ packages/common/src/Utils.ts
@@ -359,7 +359,7 @@ function isBigNumber (object: Object): boolean {
 export function toNumber (numberish: number | string | BN | BigInt): number {
   switch (typeof numberish) {
     case 'string':
-      return parseInt(numberish)
+      return parseFloat(numberish)
     case 'number':
       return numberish
     case 'bigint':
